backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
httpx
pydantic
copernicusmarine
pytest
xarray
netCDF4
numpy
scikit-learn
joblib
